requirements.txt
ibm-generative-ai
python-dotenv
genai
flask
waitress